use atol instead of atoi

+++ src/homework/work2.c
@@ -3,7 +3,7 @@
 
 int main(int argc, char **arg){
     if(argc!=2)return -1;
-    unsigned int num = atoi(*(arg+1));
+    unsigned int num = atol(*(arg+1));
     int result = 0;
     for(int i=0;i<8*sizeof(unsigned int);i++){
         if(num%2==1)result++;
