+++ src/homework/work3.c
@@ -3,7 +3,7 @@
 
 int main(int argc,char** args){
     if(argc!=2)return -1;
-    unsigned int num = atoi(*(args+1));
+    unsigned int num = atol(*(args+1));
     printf("%x\n",num);
     num = num | (1<<12);
     printf("%x\n",num);
